Rename events folder
