--- TwitchLib/Events/Services/LiveChannelMonitor/OnStreamMonitorStartedArgs.cs
@@ -1,21 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using TwitchLib.Models.Client;
-using TwitchLib.Enums;
-
-namespace TwitchLib.Events.Services.LiveStreamMonitor
-{
-    /// <summary>Class representing event args for OnChannelMonitorStarted event.</summary>
-    public class OnStreamMonitorStartedArgs : EventArgs
-    {
-        /// <summary>Event property representing channel the service is currently monitoring.</summary>
-        public List<string> Channels;
-        /// <summary>Event property representing how channels IDs are represented.</summary>
-        public StreamIdentifierType IdentifierType;
-        /// <summary>Event property representing seconds between queries to Twitch Api.</summary>
-        public int CheckIntervalSeconds;
-    }
-}
