+++ TwitchLib/Events/Services/LiveStreamMonitor/OnStreamMonitorEndedArgs.cs
@@ -1,9 +1,5 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using TwitchLib.Models.Client;
 using TwitchLib.Enums;
 
 namespace TwitchLib.Events.Services.LiveStreamMonitor
@@ -12,7 +8,7 @@ namespace TwitchLib.Events.Services.LiveStreamMonitor
     public class OnStreamMonitorEndedArgs : EventArgs
     {
         /// <summary>Event property representing channel the service is currently monitoring.</summary>
-        public List<String> Channels;
+        public List<string> Channels;
         /// <summary>Event property representing how channels IDs are represented.</summary>
         public StreamIdentifierType IdentifierType;
         /// <summary>Event property representing seconds between queries to Twitch Api.</summary>
