+++ TwitchLib/Events/Services/LiveStreamMonitor/OnStreamsSetArgs.cs
@@ -1,9 +1,5 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using TwitchLib.Models.Client;
 using TwitchLib.Enums;
 namespace TwitchLib.Events.Services.LiveStreamMonitor
 {
